refactoring; ChannelAccess, added MarkdownTextbox disabled state
