+++ webapp/i18n/en.json
@@ -2,7 +2,6 @@
   "+8G9qr": "Default text for the retrospective.",
   "+QgvjN": "Assign the owner role to",
   "+Tmpup": "You automatically receive updates when this playbook is run.",
-  "+ZIXOR": "Channel access",
   "+hddg7": "Add to run timeline",
   "/1FEJW": "ACTIVE PARTICIPANTS per day over the last 14 days",
   "/4tOwT": "Skip",
@@ -40,7 +39,6 @@
   "5CI3KH": "Contact support",
   "5FRgqE": "Downloading channel log",
   "5Ofkag": "Enable retrospective",
-  "5Ot7cd": "Determine the type of channel this playbook creates.",
   "5ciuDD": "NOT IN CHANNEL",
   "5j6GD/": "{numParticipants, plural, =0 {no participants} =1 {# participant} other {# participants}}",
   "5qBEKB": "What are playbook runs?",
@@ -322,7 +320,6 @@
   "tzMNF3": "Status",
   "u4MwUB": "Save your playbook run history",
   "uBLF+D": "What is a playbook?",
-  "uhu5aG": "Public",
   "uny3Zy": "Playbooks",
   "usa8vQ": "Send a welcome message",
   "v1DNMW": "Retrospective published by {name}",
@@ -331,7 +328,6 @@
   "v8ZnNc": "Select a team",
   "vNiZXF": "There are no runs in progress at the moment. Run a playbook to start orchestrating workflows for your team and tools.",
   "vaYTD+": "There {outstanding, plural, =1 {is # outstanding task} other {are # outstanding tasks}}. Are you sure you want to finish the run?",
-  "viXE32": "Private",
   "vjzpnC": "There are no playbooks matching those filters.",
   "vndQuC": "Slash Command Executed",
   "w0muFd": "Send outgoing webhook (One per line)",
