+++ webapp/src/components/backstage/playbook_edit/automation/channel_access.tsx
@@ -1,0 +1,126 @@
+// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
+
+import React from 'react';
+import styled from 'styled-components';
+import {useIntl} from 'react-intl';
+
+import {DraftPlaybookWithChecklist, PlaybookWithChecklist} from 'src/types/playbook';
+import {PatternedInput} from 'src/components/backstage/playbook_edit/automation/patterned_input';
+import {AutomationHeader, AutomationTitle} from 'src/components/backstage/playbook_edit/automation/styles';
+import {Toggle} from 'src/components/backstage/playbook_edit/automation/toggle';
+import {RadioInput} from 'src/components/backstage/styles';
+import {HorizontalSpacer} from 'src/components/backstage/playbook_runs/shared';
+
+interface Props {
+    playbook: DraftPlaybookWithChecklist | PlaybookWithChecklist;
+    setPlaybook: (playbook: DraftPlaybookWithChecklist | PlaybookWithChecklist) => void;
+    setChangesMade: (b: boolean) => void;
+}
+
+export const CreateAChannel = ({playbook, setPlaybook, setChangesMade}: Props) => {
+    const {formatMessage} = useIntl();
+
+    const handlePublicChange = (isPublic: boolean) => {
+        setPlaybook({
+            ...playbook,
+            create_public_playbook_run: isPublic,
+        });
+        setChangesMade(true);
+    };
+
+    const handleChannelNameTemplateChange = (channelNameTemplate: string) => {
+        setPlaybook({
+            ...playbook,
+            channel_name_template: channelNameTemplate,
+        });
+        setChangesMade(true);
+    };
+
+    return (
+        <AutomationHeader>
+            <AutomationTitle>
+                <Toggle
+                    isChecked={true}
+                    disabled={true}
+                    onChange={() => null}
+                />
+                <div>{formatMessage({defaultMessage: 'Create a channel'})}</div>
+            </AutomationTitle>
+            <HorizontalSplit>
+                <VerticalSplit>
+                    <Button onClick={() => handlePublicChange(true)}>
+                        <RadioInput
+                            type='radio'
+                            checked={playbook.create_public_playbook_run}
+                        />
+                        <Icon
+                            active={playbook.create_public_playbook_run}
+                            className={'icon-globe'}
+                        />
+                        <BigText>{'Public'}</BigText>
+                    </Button>
+                    <HorizontalSpacer size={8}/>
+                    <Button onClick={() => handlePublicChange(false)}>
+                        <RadioInput
+                            type='radio'
+                            checked={!playbook.create_public_playbook_run}
+                        />
+                        <Icon
+                            active={!playbook.create_public_playbook_run}
+                            className={'icon-lock-outline'}
+                        />
+                        <BigText>{'Private'}</BigText>
+                    </Button>
+                </VerticalSplit>
+                <PatternedInput
+                    enabled={true}
+                    input={playbook.channel_name_template}
+                    onChange={handleChannelNameTemplateChange}
+                    pattern={'[\\S][\\s\\S]*[\\S]'} // at least two non-whitespace characters
+                    placeholderText={formatMessage({defaultMessage: 'Channel name template (optional)'})}
+                    type={'text'}
+                    errorText={formatMessage({defaultMessage: 'Channel name is not valid.'})}
+                />
+            </HorizontalSplit>
+        </AutomationHeader>
+    );
+};
+
+const VerticalSplit = styled.div`
+    display: flex;
+`;
+
+const HorizontalSplit = styled.div`
+    display: block;
+    text-align: left;
+`;
+
+const Button = styled.button`
+    padding: 10px 16px;
+    border: 1px solid rgba(var(--center-channel-color-rgb), 0.16);
+    background: var(--center-channel-bg);
+    border-radius: 4px;
+    flex-grow: 1;
+    flex-basis: 0;
+    margin: 0 0 8px 0;
+
+    &:disabled {
+        background: rgba(var(--center-channel-color-rgb), 0.08);
+    }
+
+    display: flex;
+    align-items: center;
+`;
+
+const Icon = styled.i<{ active?: boolean }>`
+    font-size: 16px;
+    line-height: 16px;
+    color: ${(props) => (props.active ? 'var(--button-bg)' : 'rgba(var(--center-channel-color-rgb), 0.56)')};
+`;
+
+const BigText = styled.div`
+    font-size: 14px;
+    line-height: 20px;
+    font-weight: 400;
+`;
